+++ Commands/Branding/ConvertProvisioningTemplate.cs
@@ -81,7 +81,9 @@ namespace SharePointPnP.PowerShell.Commands.Branding
                         }
                     case XMLPnPSchemaVersion.V201505:
                         {
+#pragma warning disable CS0618 // Type or member is obsolete
                             formatter = XMLPnPSchemaFormatter.GetSpecificFormatter(XMLConstants.PROVISIONING_SCHEMA_NAMESPACE_2015_05);
+#pragma warning restore CS0618 // Type or member is obsolete
                             break;
                         }
                     case XMLPnPSchemaVersion.V201508: