+++ Commands/Properties/Resources.Designer.cs
@@ -206,6 +206,16 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Icon similar to (Icon).
+        /// </summary>
+        internal static System.Drawing.Icon pnp {
+            get {
+                object obj = ResourceManager.GetObject("pnp", resourceCulture);
+                return ((System.Drawing.Icon)(obj));
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Remove content type?.
         /// </summary>
         internal static string RemoveContentType {