+++ Samples/Modules.Install/Install-SharePointPnPPowerShellHelperModule.ps1
@@ -71,7 +71,7 @@ if ((Get-command -Module PowerShellGet).count -gt 0)
     { 
     Write-Output 'PowerShellPackageManagement now installed we will now run the next command in 10 Seconds'
     Start-Sleep -Seconds 10 
-    Install-Office365DevPNPModules -ModuleToInstall (Request-SPOOrOnPremises)
+    Install-SharePointPnPPowerShellModule -ModuleToInstall (Request-SPOOrOnPremises)
     }
     else
         {